+++ packages/main/client/static/store/adminStore.js
@@ -59,4 +59,33 @@ function AdminStore (utilStore) {
     })
   })
 
+  this.on('load_users', () => {
+    return new Promise((resolve, reject) => {
+      this.utilStore.ajaxCall({
+        method: 'get',
+        url: '../data/core/users'
+      }, true).then(users => {
+        this.trigger('users_loaded', users)
+        resolve(users)
+      }).catch(error => {
+        reject(error)
+      })
+    })
+  })
+
+  this.on('set_user_admin', (userId, isAdmin) => {
+    return new Promise((resolve, reject) => {
+      this.utilStore.ajaxCall({
+        method: 'put',
+        data: JSON.stringify({ admin: isAdmin }),
+        url: '../data/core/users/' + userId + '/admin'
+      }, true).then(data => {
+        this.trigger('user_admin_changed', data)
+        resolve(data)
+      }).catch(error => {
+        reject(error)
+      })
+    })
+  })
+
 }